refactor(codec-video): extract composition modules + ProcessRunner (closes #327)
Per the AI-shape audit (docs/research/2026-05-13-ai-shape-audit.md,
merged via #328), packages/codec-video/src/hyperframes-wrapper.ts (409
lines) carried two canonical AI-code smells:

1. buildHyperFramesHtml() — 168-line monolithic HTML builder with two
   nearly-identical branches (66% duplication between SVG-slide and
   scene-card paths). Inline CSS strings were duplicated.
2. spawnProcess() — 58-line subprocess wrapper tangling timeout +
   signal handling with stdout/stderr buffering and structured-error
   extraction.

This refactor splits along both seams.

Three composition files:

- packages/codec-video/src/compositions/svg-slide.ts (new, 117L) —
  buildSvgSlideHtml() for the inline-SVG composition path; carries its
  own resolveSlideDurations + stripXmlDeclaration helpers (the BOM
  regex preserves the  escape).
- packages/codec-video/src/compositions/scene-card.ts (new, 132L) —
  buildSceneCardHtml() for the auto-scene composition path; carries
  its own sceneTone helper.
- packages/codec-video/src/compositions/shared.ts (new, 29L) —
  STAGE_WIDTH / STAGE_HEIGHT constants, sanitizeCompositionId,
  escapeHtml, and the BASE_CSS rule both compositions extend.

One subprocess helper:

- packages/codec-video/src/process-runner.ts (new, 96L) — runProcess()
  with typed timeout + stdout/stderr capture. Kept inside codec-video
  for now; will lift to packages/core/src/utils/ in a follow-up if a
  second codec needs the same shape (likely once audio's Kokoro
  wiring is touched). errorPrefix parameter preserves
  domain-specific error phrasing; timeoutHint preserves the
  operational hint about WITTGENSTEIN_HYPERFRAMES_RENDER_TIMEOUT_MS.

Updated hyperframes-wrapper.ts (409 → 167L) is now a thin orchestrator:
renderWithHyperFrames + runHyperframesRenderToMp4 + snapHyperframesFps +
resolveHyperFramesArtifactPath + a two-line buildHyperFramesHtml
dispatcher that delegates to the right composition module based on
whether the composition carries inlineSvgs.

What did NOT change:

- Public API. renderWithHyperFrames keeps the same signature; the
  index.ts re-export (export * from "./hyperframes-wrapper.js") is
  unaffected.
- HTML output semantics. Per-composition CSS is semantically identical
  to the original — the BASE_CSS extraction means the resulting HTML
  has two html/body rules instead of one combined, but CSS cascade
  merges them identically. Existing tests use toContain checks (not
  byte-equality) and pass.
- MP4 encoding behavior. npx hyperframes render args, env vars,
  timeout, FPS snapping all preserved.
- Error messages. errorPrefix + timeoutHint preserve the original
  domain-specific phrasing for hyperframes render timeouts and
  failures.

Validation:

- pnpm --filter @wittgenstein/codec-video typecheck — clean.
- pnpm --filter @wittgenstein/codec-video test — all 5 tests pass
  (3 codec tests + 2 playable-slideshow-html tests).
- pnpm --filter @wittgenstein/codec-video lint — 0 errors, 0 warnings.

hyperframes-wrapper.ts shrinks 409 → 167 lines (59% reduction, under
the audit's "under 200 lines" target). Total LOC across the 5 files
is 541 (vs 409 original) due to file headers, imports, and per-module
JSDoc — expected cost of the seam.

Refs:
- Closes #327.
- Audit source: docs/research/2026-05-13-ai-shape-audit.md (merged via
  #328).
- Sibling refactor PRs (both merged): #337 (landscape renderer), #338
  (sensor operator strategies + LoupeRenderer).
- Related video work: #282 HyperFrames distillation, #265 HF receipts
  — both downstream of this refactor; cleaning the wrapper shape now
  reduces friction when those land.
